Fix logic for changing priority to first entry
Need parentheses to allow if condition to run. Only first #priority query is
selected with query selector. So, its value is wrongly set. For example,
when div with priority medium/low is clicked, populateForm assigns the
value medium/low to button corresponding to high. So, selecting high
after does not have an effect!